+++ changelog.md
@@ -73,6 +73,7 @@ Codion Change Log
 - FilterTable.Builder.Columns added, replacing the columns parameter in FilterTable.builder().
 - FilterTable.Builder.Model added, replacing the tableModel parameter in FilterTable.builder().
 - FileSelectionDialogBuilder.fileFilter() renamed filter().
+- ProgressWorkerDialogBuilder.BuilderFactory added, replacing overloaded methods in Dialogs.
 ### is.codion.swing.framework.ui
 - ApplicationPreferences.fromString() bug fixed, FRAME_MAXIMIZED_KEY now used.
 - EntityComboBoxPanel.Builder.EditPanel added, replacing the editPanel parameter in EntityComboBoxPanel.builder().